This was showing up as a small hotspot (around 20% CPU time for get requests) in replicator2 read. Seemed like an easy fix.

```
benchmark                            old ns/op     new ns/op     delta
BenchmarkEncodeRevisions/empty-8     231           239           +3.46%
BenchmarkEncodeRevisions/1-8         1145          297           -74.06%
BenchmarkEncodeRevisions/3-8         2043          351           -82.82%
BenchmarkEncodeRevisions/26-8        20833         909           -95.64%
BenchmarkEncodeRevisions/1000-8      1069939       22683         -97.88%

benchmark                            old allocs     new allocs     delta
BenchmarkEncodeRevisions/empty-8     3              3              +0.00%
BenchmarkEncodeRevisions/1-8         10             5              -50.00%
BenchmarkEncodeRevisions/3-8         15             5              -66.67%
BenchmarkEncodeRevisions/26-8        126            5              -96.03%
BenchmarkEncodeRevisions/1000-8      6002           5              -99.92%

benchmark                            old bytes     new bytes     delta
BenchmarkEncodeRevisions/empty-8     368           368           +0.00%
BenchmarkEncodeRevisions/1-8         488           392           -19.67%
BenchmarkEncodeRevisions/3-8         600           408           -32.00%
BenchmarkEncodeRevisions/26-8        3217          792           -75.38%
BenchmarkEncodeRevisions/1000-8      120109        16760         -86.05%
```